requirements.txt
affinegap==1.10
astroid==2.1.0
backcall==0.1.0
bleach==3.1.0
BTrees==4.5.1
categorical-distance==1.9
Click==7.0
colorama==0.4.1
cycler==0.10.0
datetime-distance==0.1.3
decorator==4.3.2
dedupe==1.9.5
dedupe-hcluster==0.3.6
dedupe-variable-datetime==0.1.5
defusedxml==0.5.0
DoubleMetaphone==0.1
entrypoints==0.3
fastcluster==1.1.24
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
future==0.17.1
haversine==2.0.0
highered==0.2.1
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Levenshtein-search==1.4.4
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
patsy==0.5.1
persistent==4.4.3
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.8
Pygments==2.3.1
pyhacrf-datamade==0.2.3
PyLBFGS==0.2.0.12
pylint==2.2.2
pyodbc==4.0.26
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2018.9
pywinpty==0.5.5
pyzmq==18.0.0
qtconsole==4.4.3
rlr==2.4.5
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplecosine==1.2
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.3.1
statsmodels==0.9.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
Unidecode==1.0.23
virtualenv==16.2.0
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.0
widgetsnbextension==3.4.2
wrapt==1.11.1
zope.index==4.4.0
zope.interface==4.6.0
